fix: zkCompile output
